pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "psutil>=7.0.0",
    "pycryptodome>=3.23.0",
    "pykeepass>=4.1.1.post1",
]
